single test file hotfix
bring back both tests under one file to prevent double block registration
